Allow per-realm and per-datacenter configuration
Read the realm (labs or production) from /etc/wikimedia-realm and the
datacenter (pmtpa or eqiad) from /etc/wikimedia-site, and change the
various switches to switch based on $realm instead of cluster (which was
basically $realm trying to pretend to be $datacenter somehow).

This also expands the realm-specific configuration switching to include
files such as all.dblist and wikiversions.dat, allowing us to remove the
live hacking of those files from the Beta config.

Bug: 39082
Bug: 41133

Change-Id: I00390aacbec083bff8219e41a4f9f4fe9cfdca1d