Wikipedia archive import script
This script was deleted from the core in r100128, only a year after I
wrote it (r78674). I would like it to be in version control of some
description, so that development can continue, and so that it can more
easily be found by researchers. Maybe it will live longer in this
extension.

Change-Id: Ia6f209e06d3a7a616c85a0b6d369e31a35d813f6