Add language override and extra link support for dumpInterwiki
This will allow chapter wikis to have the project interwiki prefixes (e.g.
"w:") point to the relevant projects for that chapter.

I have also unified a lot of the duplicated config variables in both
dump and rebuild, it now all lives in dump. But I don't think anyone uses
rebuild anymore, it can probably just be deleted.

Note that this will change interwikis for several chapter wikis without
them having specifically asked for it (I just copied the list out of
InitialiseSettings). But I don't think this is a problem - rather, it is a
Good Thing. For instance, the Argentinian chapter currently gets
interproject links to Arabic projects, and I don't think they'll be
sad to lose those!

Bug: 56207
Bug: 55733
Bug: 53011
Change-Id: Ie327c77eaf6f9b65a3bb6b8741f88d60fabfdae6