Add a script to clean up page_props
In some cases such as when a deletion by property value is needed,
runBatchedQuery.php is insufficient and a custom solution is needed.
The first added cleanup condition concerns values added by MobileFrontend.
The amount of them is serious enough for large wikis to warrant a cleanup.
However, I want to create an universal script that can be used for
multiple cleanups rather than a one-off extension-specific script.

Change-Id: Ifd4eeda9e5c8b4c9e15ac88978e5e3e238111589