Avoid to hardcode /srv/mediawiki
On the Wikimedia Cluster, this script is called from mwscript,
which uses the MEDIAWIKI_DEPLOYMENT_DIR environment variable to
represent the /srv/mediawiki path.

We now use also this environment variable to allow this script
to be used on machines with path another configuration, for example
to allow interwiki map update from a developer machine.

Change-Id: Icb8a1d7ffea739b6cfe631edebe00ca70127972d