blameStartupRegistry.php: Add total count/bytes
The current Grafana dashboard tries to approximate this using
sumSeries() over all the components.

This is problematic because each component count is sent separately
to Graphite, which means they don't all arrive at once.

Then, when you query for the last 7 days, the beginning of the graph
often looks like there was a regression because the first minute
of the graph is a sum over only some of the components (lacking the
ones that were flushed on the other side of the minute boundary).

Instead of relying that at run-time, just send the totals explicitly.

Change-Id: Icc8961b13182acfd81939685d80a9437537b5795