addWiki.php: Redefine RevisionStore service for the wiki being created
The RevisionStore service provides a high level interface for accessing
and manipulating page revisions. Internally, it ends up calling ExternalStoreDB
via SqlBlobStore -> ExternalStoreAccess -> ExternalStoreFactory; ExternalStoreDB
then handles inserting raw revision content into external storage.

By default, the RevisionStore service is created by calling
RevisionStoreFactory::getRevisionStore without passing an explicit database domain,
to operate in the context of the local wiki. This causes an implicit database domain
of false to be passed down the call stack to ExternalStoreDB.

At this point, due to T200471 backwards compatibility requirements, ExternalStoreDB
inspects the server array of the LoadBalancer instance for the external DB cluster,
and uses the 'dbname' value from the server array when establishing a connection.
This ignores the effect of any previous redefineLocalDomain() calls on the LoadBalancer
instance, as LoadBalancer::redefineLocalDomain() does not update the server array.
In the context of addWiki.php, this causes the connection to use the DB name of the wiki
the maintenance script is being ran on (aawiki), rather than the DB name of the newly created
wiki. Thus, revision content is inserted into the wrong database on the external cluster,
causing pages to show broken content on the new wiki.

This patch makes addWiki.php redefine the RevisionStore service to be created by calling
RevisionStoreFactory::getRevisionStore with the DB name of the newly created wiki. This
causes the explicit database domain of the new DB name to be passed down the call stack
to ExternalStoreDB. ExternalStoreDB will then use the explicit DB name when establishing
database connections.

Bug: T212881
Change-Id: I46a3f7cca33a542e0d959aa64b5e21033d8db7ca