Remove CirrusSearch initialization from addWiki.php
addWiki.php is run in the context of a dummy wiki that selects for the
correct database shard. CirrusSearch is also sharded, but by a
different metric of the wiki id. This conflict mostly means search
needs to be initialized in a separate maintenance script execution and
not as a child of addWiki.php.  The cluster selection / for loop will
be moved into CirrusSearch to provide a single invocation that will
perform the full operation, and documentation will be updated to
require performing that invocation.

Bug: T254331
Change-Id: Ibca011525ad0d3a80feb87db73355cb28abee91a