blameStartupRegistry: Call StartupModule::getScript instead of hardcoding
Follows-up 492d91175e9dc, which moves most of the mediawiki.js content
to a new file, which is not yet monitored here.

Before (with mediawiki.loader.js manually added, or 492d91175e reverted):

```
| Component | Modules | Bytes
|-- |-- |--
| MediaWiki core       |   231 |    9,570
| startup_js           |     0 |    7,001
| Vector               |     6 |      209
```

After (latest master of core + this patch):

```
| Component | Modules | Bytes
|-- |-- |--
| MediaWiki core       |   231 |    9,570
| startup_js           |     0 |    7,050
| Vector               |     6 |      209
```

It is slightly larger, as it now correctly measures some of the
$CODE and $VARS values. I originally made these fixed nulls here
instead of real values. I think that is from a time where we ran this
on one wiki only and I didn't want to bias these toward any given
wiki's configuration. But, we now run this script on many more wikis,
so this is no longer a concern.

The difference (before minification) is saved at
<https://phabricator.wikimedia.org/P17134>.

Bug: T290213
Change-Id: I63bf1922af593b7a144ef5f6d036f9a5e23cec09