maintenance: Add email search script
This script takes as input a users email and produces a json blob
indicating the username, email and email verification date (if present)
for all users matching the email address.

The intention is to call this script remotely using a cookbook and have
that cookbook process the json response in combination with data from
gsuite and possibly namely.

Usage:

$ getUserByEmail.php --email='user@example.org'
[
  {
    "username": "My_user",
    "email": "user@example.org",
    "email_authenticated_date": "20151231165901",
  }
]

Bug: T259746
Change-Id: If8f7a0da09b51af3bb30396d7d7f126dd44ad501