blameStartupRegistry: Fix clash in $startupBytes variable name
Follows-up Ia2816b6af73, which renamed $totalBytes to $startupBytes
(to differentiate it from $contentBytes in I14d212011d41f1144e), but
I accidentally renamed $bytes also to $startupBytes.

The result was that all wikis had their total size reported as only
the size of the last extension that we check + the size of startup.js,
thus throwing away the running total.

Bug: T295413
Change-Id: Ib63ec06ccfb969219289d3d72820259b36a3cffd