blameStartupRegistry: Escape dots in stats component name
This issue has been around for a long time apparently. I only noticed
now that Grafana was showing an extension called "Wikidata" in the
component list, with a module called "org" and no metrics.

Turns out this is actually from the "Wikidata.org" extension.
The dot is not just in its display name but also its internal
repository name...

Bug: T295413
Change-Id: I41305a1df8d43ee9790113bc560aae9768ccca20