Stop excluding the 'wikipedia' interwiki prefix
This code snipped added the '_wiki:wikipedia' => '' entry to the
interwikimap. Since it maps to an empty string, it was never treated
as a valid entry within core.

With mediawiki/core I736cb74efc267fd2473a3267471735238217251c, the
behaviour has changed and is now treating empty string as valid
interwiki prefixes. As a result the 'wikipedia' interwiki is broken on
multiple wikis.

Also we were not excluding other prefixes which did the same thing
in the past, like 'wikipediawikipedia'.

Bug: T301936
Change-Id: I06bf2fe302edeb15f38a0f59c056acedebdbed44