renameWiki: Remove pointless tableNames() call
IDatabase::tableNames() has *always*, ever since it was introduced in
MediaWiki core commit ac549401d4 way back in 2004, returned prefixed
table names *for the names it was given*, like so:

    $db->tableNames( 'user', 'page' )
    // [ 'user' => 'prefix_user', 'page' => 'prefix_page' ]

Calling it with an empty argument list will produce an empty array. It
is not a method that will return the names of all MediaWiki core tables,
and as far as I’m aware no such method exists. I can only conclude that
this whole block of code, added nine years ago in change I1b9fa6bc7d
(commit d794af188c), never did anything. Replace it with a note
explaining that (non-blob?) tables aren’t being renamed (which might not
be desirable anyways, or at least the latest rename for be-tarask did
not rename the database away from be_x_oldwiki); also, add a reference
to a relevant task at the top of the file.

The direct motivation for this is simply to remove the call to
tableNames() so that we can start to deprecate the method in core.

This reverts commit d794af188cda839c21145a70b02fb5cd19c40ba1.

Bug: T48141
Change-Id: I0ff249b620e214aedbb659b1675f26b7bce38e28