blameStartupRegistry: Fix "Notice: Undefined key startupBytes"
Line 210 and 217 access 'startupBytes' on each component, which
requires the field to exist. Virually all components have their
array key in $startupBreakdown created during the foreach loop
where this is created correctly.

The placeholder "unknown" component, however, is created upfront
and was missing this key (or rather, it named it something else).

This isn't seen in production, since there is at least one unknown
module and thus the foreach loop auto-corrects it on the first
iteration.

This bug can be reproduced locally by harcoding `$moduleNames = [];`
on line 55 and then running `php blameStartupRegistry.php`.

Change-Id: I0343fd1d4a4407920ea7bb0effdd2c27ad8d5aa7