blameStartupRegistry: migrate off string-assumed Module::getScript()
In commit 93a27630c (I7017de5757b22263f1), even StartupModule,
which was one the last modules to still return a string from getScript,
now returns an array.

Call ResourceLoader->makeModuleResponse() instead, as the rest of this
script already does for all other modules.

For it to produce the same result, two other changes are needed:

* setName(), to avoid a PHP error for invalid null $name.
* setRaw(true), to match the real startup request and thus avoid
  appending `mw.loader.state({"startup":"ready"});` to the output
  as an only=scripts request would otherwise do.

Locally, this results in byte-identical response. Both before and
after, the script measures exactly:

> startup_js: 6,680 B

Bug: T346800
Change-Id: I83879ebcf97695ae16c322f8f8a744c49039fa0e