Remove defgroup/ingroup tags from file/class doc blocks
The "Maintenance" Doxygen group is defined in MediaWiki core, which
these aren't part of. And given there are virtually no other files
in this repo, it doesn't seems like a useful group to define
locally, either.

The "Wikimedia" Doxygen is undefined, except for in one random file,
and similary has no usage or purpose. Note that this repo does not
even have a Doxygen publishing build configured in CI.

Change-Id: I9cffd2871502aa539e8fd9a7bdb53107bc6cd0b8