Reenable phan
Almost all scripts in this repo are in the root directory. So in order
for phan to take a look at them, we would have to add '.' to the config.

This was initially the case, but it got removed in b1a35af516fb.

But doing that is a bit clunky and also causes other issues, so its
easier to just move the scripts to a proper subfolder.

Change-Id: Ice7fa4b4a7a59c6245a4da49d374436caa1a5c94